refactor/#470 : 기존 DailyMeal 도메인의 컬렉션 타입을 MealFoodEntity로 변경 및 DailyMeal을 DailyMealFoods로 클래스명 수정
